refactor: consolidate events service
